requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
pinecone==5.3.1
pydantic==2.9.2
openai==1.51.2
httpx<0.28
requests==2.32.3
# Optional: airbyte-api for SDK-based connection create (first variant tried)
airbyte-api>=0.53.0
